refactor: Cleanup css generation
